type must be 'int' since there is no 'customer_entity_smallint' table

+++ app/code/community/Americaneagle/Visual/sql/americaneagle_visual_setup/upgrade-1.0.3-1.0.4.php
@@ -10,7 +10,7 @@ $installer = $this;
 $installer->startSetup();
 
 $installer->addAttribute('customer', 'tax_exempt', array(
-    'type' => 'smallint',
+    'type' => 'int',
     'input' => 'boolean',
     'label' => 'Tax Exempt',
     'global' => Mage_Catalog_Model_Resource_Eav_Attribute::SCOPE_GLOBAL,
